feat: change controller.js to controller.ts
